decodeConfirmTx
